- updated Readme file; - refactor user menu buttons
